Dev: unittest: Adjust unit test code for previous changes
